Post-defence reviews and modifications
Updated data

In file "ajar_scores.csv":
* Added new data (from late experiments)
* Replaced "diff2" with "ratio"

Update .gitignore

* Ignore ".Rprofile"
This file is used to automatically apply a config option for knitrdrawio that only makes sense on my (local) machine, not on a CI or another machine.

* Ignore "__pycache__" folders. They are artifacts from Python.

Better Makefile

* Added ability to self-document (using make-help).
* Added documentation (comments).
* Added variables that control where to build the book.
* Removed unnecessary code.

Added list of publications

A new appendix contains the list of my publications, as a definition list.

Added positioning wrt value alignment problem

Added a small paragraph and 2 references that position our ethical injection w.r.t. the value alignment problem in Chapter 3.

Added more references on RL and human behaviours

Added 2 sentences and 2 refs in Chapter 2.

Added positioning wrt VAF

Added a remark paragraph and 1 reference to position our work (AFJD) with Value-based Argumentation Frameworks in Chapter 5.

Described algorithms line-by-line

* Some paragraphs were modified to add references to the line number of algorithms.
* Paragraphs have been added for other algorithms, which describe line-by-line how the algorithm works.

Added discussion on extending implementation to other domains

Added definition and references for Smart Grids

Added List of Algorithms

We use "algorithms2e" for algorithms, which provide the "listofalgorithms" command. Other commands such as "lstlistoflistings" do not work.

Producing the list of algorithms creates a "thesis.loa" file, that we now ignore.

Fixed the link to appendix

We have created a "false" Appendix part to make the table of contents clearer. However, since it is not a real part, hyperref pointed to the previous section (References).
We now use a phantomsection to get hyperref to point to the correct place, i.e., when appendices begin.

Updated the notes file to remove some old TODO, and added a link to better understand phantomsection.

Removed old (useless) code

In "index.Rmd":
* Removed commented configuration for Table of Contents, Tables, Figures, Listings that are not used.

In "template.tex":
* Removed commented code for the jury members (was hardcoded instead to simplify).
* Removed commented code for the list of department directors (not used).
* Removed commented code from an old template.
* Removed commented variable not used (thesis version).
* Replaced the PDF subject "Documentation" by the thesis title.
* Removed commented code about natbib.
* Removed commented code about German translations.
* Removed code to change the name of "List of Listings" (we do not use this code, as we use algorithm2e).
* Removed old commented code from an old template, about the ToC, LoT, LoF, and abstract.

Fixed tables captions

Some tables' captions included Markdown code that was not parsed correctly, resulting in an incorrect caption, e.g., "Scores for the *annual* scenario", instead of an italicized "annual".
This was solved by replacing the `tab.cap` values by references to a Markdown paragraph.

Deleted unused file

The file "titlepages.tex" was supposed to hold the title page in a separate file, to better separate the template from the cover.
In the end, it did not work and we simply harcoded the cover inside the template.

Better list of publications

The previous definitions list allowed for the definition to be on the same line as the label, which was not ideal in this case. Added a LaTeX block to put the definition on the next line.
Also modified the definitions to use a "non-compact" format, and avoid using the "\" blank-space hack.

Added discussion paragraph on ethical thresholds

The paragraph describes the use of disjunctions of conjunctions instead of a single vector of thresholds.

Improved descriptions of diagrams

Added or modified paragraphs associated to a diagram figure, to better explain what can be seen in the diagram, and how to read it.

Distinguished definitions from literature

Some definitions are taken from literature, but this point was not very clear.
Some phrases and references have been added to these definitions to better explicit where they come from.
Definitions without such phrase or reference are original.

Also, removed a duplicate citation key (dung).

Improved Hybrid Neural-Symbolic section in SOTA

Added multiple references, including to some of the earliest works, examples in RL, surveys, and positioning works.
Added several paragraphs to discuss these references.

Added discussion on leximin

In Chapter 5, the "leximin" variants could be used as aggregation functions.
Added references to (computational) social choice, and a paragraph in the Discussion section.

Added table of MARL approaches

In the SOTA, added a table listing the approaches that we retain from the survey by Hernandez-Leal et al.
This makes it clearer to describe how many approaches require local observations or additional data.
Also added a reference to the 2 approaches that require only local observations.

Note: one of the approaches include a math symbol. Rendering math within flextable (for both HTML and PDF output) would not be convenient in this case. Directly using the Unicode character is perfectly handled in HTML; for PDF, it requires setting up the character in "template.tex".

fixup! [WIP] Added table of MARL approaches

Multiple reviews

* Introduction: replaced "gaming" with "specification gaming" in the objectives, and added a footnote with a link for more details about what it entails.

* Learning (Algo 4.2): added description of data, and fixed typos (wrong symbols)

* Learning: replaced "our experiments proved" with "our experiments highlight"

* Judgments (Def 5.1): Clarified the "aliveness condition function", and the attack relation.

* Judgments (Experiments): changed the definition of the "ratio" function to make it clear that we update a global counter at t based on the previous value at t-1.

* Judgments (Discussion): added reference and details for the Mycin example.

* Dilemmas (bellman morl equation): added a small description of the "k" iterator.

* Replaced occurrences of "taking a decision" (or similar) with "making a decision", which is more correct.
This refers to the "decision-making" process rather than the fast (not reflected) "decision-taking".

* Dilemmas: changed two equations to avoid them overflowing into the margin.
One of them was simplified; the other was split into 2 inlines, and we added the LaTeX "sloppy" command.

Fixed inline reference

More details on Smith's work

* Added a justification on why we chose Smith's work as a basis for our RL algorithms.
* Added a remark paragraph listing our differences with Smith's original work.

Details on the world's state containing actions

In the argumentation-based judgments, the world state also contains the agents' actions.
Added details to precise that this makes a limited history, what it entails, and how we can expand it.

Improved STS

Added 2 paragraphs and 2 references on Socio-Technical Systems (STS).