PDF and LaTeX improvements
Disable the scrhack "setspace" hack

Package scrhack tries to improve other packages by "hacking" them.
This produced a warning when trying to "hack" the setspace package (definitions `\onehalfspacing` and `\doublespacing` not found).
We now disable this specific hack, to prevent these warnings.

Better placement of algorithms in Ch6

The algorithms in Chapter 6 (dilemmas) are a bit long, and, due to how LaTeX tries to place floats and text on a page, they had problems.
Either they would appear too late in the manuscript, or they would appear on a whole page, with a single line directly above or after them...
We now place a bunch of `\clearpage` to force LaTeX to flush everything before and after them. This makes sure that they appear exactly where we want them, on a whole page, and that text is not intertwined with them. Kind of a hack, but it works.

Removed warnings about figure placement

For some large figures, we used the float specifier `!h` to avoid them being placed too far away.
LaTeX produces warnings indicated that these specifiers were changed to `!ht` instead.
We now directly use `!ht` to avoid these warnings, the result seems to be the same.

Update ToC

* Made sure that hyperlinks point to the correct place, by using "\cleardoublepage" and "\phantomsection".
* Changed the format of the "List of Figures / Tables / Algorithms" in the ToC from chapter to section. This required removing default toc ("notoc" in scrreprt) and adding ourselves the toc lines ("addcontentsline").

Note: tocbasic complains about the "number width of figure toc entries", caused by entries spawning on multiple lines. This should not be a problem, and I am not sure how to solve.

Fixed ToC numbering

In the ToC, the list of figures / tables / algorithms had the wrong page numbers (the hyperref still pointed to the correct page).
Solved this by changing (incrementing) the number of LaTeX compilation, so that LaTeX can first build the ToC, then see how much space it needs, and re-build the ToC with correct page numbers.
This also seems to remove the warnings about number width.

Replaced math code

In some places, replaced the "hardcoded" math symbols, e.g., "\mathbb{R}" with our custom definitions, e.g., "\RR".