Added scripts
These scripts were used to process experiments data and to produce the resulting data files used in the manuscript.
However, they cannot be used as-is, because they lack the original experiments data, which are far too vast (several tens of GBs) to provide.

Removed unnecessary code in "scripts.smartgrid_load_profiles.py"